9C-279 - Add Filter to DeviceProcess getting Apps

+++ modules/repository/src/main/scala/com/fortysevendeg/ninecardslauncher/repository/model/Model.scala
@@ -11,8 +11,8 @@ case class AppData(
   category: String,
   imagePath: String,
   colorPrimary: String,
-  dateInstalled: Double,
-  dateUpdate: Double,
+  dateInstalled: Long,
+  dateUpdate: Long,
   version: String,
   installedFromGooglePlay: Boolean)
